Simplify PatchDriver and CLI

+++ src/mass_driver/drivers/counter.py
@@ -13,9 +13,9 @@ class Counter(PatchDriver):
     target_count: int
     counter_file: Path
 
-    def detect(self, repo_path: Path) -> bool:
-        """Detect if we need to patch the counter file"""
-        counter_filepath_abs = repo_path / self.counter_file
+    def run(self, repo: Path, dry_run: bool = True) -> bool:
+        """Process the counter file"""
+        counter_filepath_abs = repo / self.counter_file
         if not counter_filepath_abs.is_file():
             print("File not found: let's Patch!")
             return True
@@ -28,9 +28,6 @@ class Counter(PatchDriver):
         print(
             f"Measured: {counter_number}, target: {self.target_count}. Different? {counter_is_different}"
         )
+        if counter_is_different and not dry_run:
+            counter_filepath_abs.write_text(str(self.target_count))
         return counter_is_different
-
-    def patch(self, repo_path: Path):
-        """Increment the counter"""
-        counter_filepath_abs = repo_path / self.counter_file
-        counter_filepath_abs.write_text(str(self.target_count))
